Move UserDisplay to common components
